extract loader_builder module
